[FIX] Cleaning up the task and command handling classes.
