+++ src/main/java/com/squarepolka/readyci/tasks/app/android/AndroidCreateApkFile.java
@@ -2,8 +2,6 @@ package com.squarepolka.readyci.tasks.app.android;
 
 import com.squarepolka.readyci.taskrunner.BuildEnvironment;
 import com.squarepolka.readyci.tasks.Task;
-import com.squarepolka.readyci.tasks.TaskExecuteException;
-import com.squarepolka.readyci.util.PropertyMissingException;
 import org.springframework.stereotype.Component;
 
 @Component
