+++ src/main/java/com/squarepolka/readyci/tasks/readyci/TaskExecuteException.java
@@ -1,4 +1,4 @@
-package com.squarepolka.readyci.tasks;
+package com.squarepolka.readyci.tasks.readyci;
 
 public class TaskExecuteException extends RuntimeException {
